Upgrade program.
- Add poetry;
- Add typing to functions.
