+++ deck.json
@@ -1,4 +1,86 @@
-["Ashnod's Altar",
-"Skullclamp",
-"Spawning Pit",
-"Sol Ring"]
+["Abrade",
+"An Offer You Can't Refuse",
+"Anger",
+"Animate Dead",
+"Arcane Denial",
+"Arcane Signet",
+"At Knifepoint",
+"Baral, Chief of Compliance",
+"Bloodthirsty Blade",
+"Bojuka Bog",
+"Chainer, Dementia Master",
+"Chaos Warp",
+"Codex Shredder",
+"Command Tower",
+"Crumbling Necropolis",
+"Diluvian Primordial",
+"Dimir Signet",
+"Dismember",
+"Doomed Necromancer",
+"Dread Return",
+"Duelist of the Mind",
+"Eroded Canyon",
+"Etali, Primal Storm",
+"Evolving Wilds",
+"Fabled Passage",
+"Feed the Swarm",
+"Ghitu War Cry",
+"Ghost Vacuum",
+"Gisa, the Hellraiser",
+"Goblin Electromancer",
+"Hullbreaker Horror",
+"Iridescent Vinelasher",
+"Izzet Signet",
+"Jagged Barrens",
+"Jin-Gitaxias, Core Augur",
+"Laughing Jasper Flint",
+"Lazav, Familiar Stranger",
+"Lightning Bolt",
+"Maestros Theater",
+"Magda, the Hoardmaster",
+"Mayhem Devil",
+"Molten Primordial",
+"Mystic Remora",
+"Nezahal, Primal Tide",
+"Nightscape Familiar",
+"Niv-Mizzet, Parun",
+"Obsessive Stitcher",
+"Peek",
+"Pongify",
+"Rakdos Charm",
+"Rakdos Signet",
+"Rapid Hybridization",
+"Raven of Fell Omens",
+"Reality Shift",
+"Reliquary Tower",
+"Rhystic Study",
+"Sepulchral Primordial",
+"Shipwreck Marsh",
+"Shivan Reef",
+"Siphon Insight",
+"Smoldering Marsh",
+"Snap",
+"Soured Springs",
+"Spatial Binding",
+"Steam Vents",
+"Stitch Together",
+"Storm-Kiln Artist",
+"Stormcarved Coast",
+"Stormcatch Mentor",
+"Sulfur Falls",
+"Sulfurous Springs",
+"Sunken Hollow",
+"Talisman of Dominance",
+"Talisman of Indulgence",
+"Temple of the False God",
+"Terramorphic Expanse",
+"Thought Scour",
+"Tinybones Joins Up",
+"Unable to Scream",
+"Underground River",
+"Urborg, Tomb of Yawgmoth",
+"Vadmir, New Blood",
+"Vandalblast",
+"Victimize",
+"Wonder",
+"Marchesa, Dealer of Death"]
