+++ pyproject.toml
@@ -1,0 +1,20 @@
+[tool.poetry]
+name = "fair-commander"
+version = "0.1.0"
+description = ""
+authors = ["Marco Ludwig <marcoludd@gmail.com>"]
+readme = "README.md"
+
+[tool.poetry.dependencies]
+python = "^3.10"
+
+
+[tool.poetry.group.dev.dependencies]
+black = "^24.10.0"
+mypy = "^1.12.0"
+pylint = "^3.3.1"
+pytest = "^8.3.3"
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
